+++ runtime/basis/Posix/Process/execp.c
@@ -6,13 +6,13 @@ C_Errno_t(C_Int_t) Posix_Process_execp (NullString8_t fNStr,
                                         Vector(C_Size_t) aOff) {
   const char      *file;
   char            **args;
-  int             aLen;
+  uintmax_t       aLen;
   int             res;
 
   file = (const char *) fNStr;
   args = (char **) aPtr;
   aLen = GC_getArrayLength((pointer)aPtr);
-  for (int i = 0; i < aLen - 1; i++) {
+  for (unsigned int i = 0; i < aLen - 1; i++) {
     args[i] = (char *)aStr + ((size_t*)aOff)[i];
   }
   args[aLen - 1] = NULL;
